[fact] Build intermediate tree from change set #96
